+++ internal/varlogctl/controller.go
@@ -2,13 +2,13 @@ package varlogctl
 
 import (
 	"context"
+	"encoding/json"
 	"io"
 
-	"github.com/kakao/varlog/internal/varlogctl/result"
 	"github.com/kakao/varlog/pkg/varlog"
 )
 
-type ExecuteFunc func(context.Context, varlog.Admin) *result.Result
+type ExecuteFunc func(context.Context, varlog.Admin) (any, error)
 
 type Controller struct {
 	config
@@ -23,10 +23,21 @@ func New(opts ...Option) (*Controller, error) {
 	return ctl, nil
 }
 
-func (c *Controller) Execute(ctx context.Context) (res *result.Result) {
+func (c *Controller) Execute(ctx context.Context) (any, error) {
 	return c.executeFunc(ctx, c.admin)
 }
 
-func (c *Controller) Print(res *result.Result, writer io.Writer) error {
-	return Print(res, c.pretty, writer)
+func (c *Controller) Decode(res any) ([]byte, error) {
+	if err, ok := res.(error); ok {
+		panic(err)
+	}
+	if c.pretty {
+		return json.MarshalIndent(res, "", "\t")
+	}
+	return json.Marshal(res)
+}
+
+func (c *Controller) Print(buf []byte, writer io.Writer) error {
+	_, err := writer.Write(buf)
+	return err
 }